+++ sdk/examples/utils/Utils.h
@@ -54,8 +54,8 @@ class Utils {
     PrintIndexParam(const milvus::IndexParam& index_param);
 
     static void
-    BuildEntities(int64_t from, int64_t to, std::vector<milvus::VectorData>& entity_array, std::vector<int64_t>& entity_ids,
-                  int64_t dimension);
+    BuildEntities(int64_t from, int64_t to, std::vector<milvus::VectorData>& entity_array,
+                  std::vector<int64_t>& entity_ids, int64_t dimension);
 
     static void
     PrintSearchResult(const std::vector<std::pair<int64_t, milvus::VectorData>>& entity_array,
@@ -68,7 +68,7 @@ class Utils {
     static void
     DoSearch(std::shared_ptr<milvus::Connection> conn, const std::string& collection_name,
              const std::vector<std::string>& partition_tags, int64_t top_k, int64_t nprobe,
-             const std::vector<std::pair<int64_t, milvus::VectorData>>& entity_array,
+             std::vector<std::pair<int64_t, milvus::VectorData>> search_entity_array,
              milvus::TopKQueryResult& topk_query_result);
 
     static void